feat/#261/CodingUsAnswerCRUD create/delete/select

+++ cu-project/src/components/units/codingus/question/detail/QnADetail.Container.tsx
@@ -7,6 +7,8 @@ import {
   IMutationDeleteStackArgs,
   IMutationUpdateStackArgs,
   IQuery,
+  IQueryFetchStackCommentorderbycreateArgs,
+  IQueryFetchStackCommentorderbylikeArgs,
   IQueryFetchStackOnebystackidArgs,
 } from "../../../../../commons/types/generated/types";
 import CodingUsLayout from "../../layout/CodingUsLayout";
@@ -14,12 +16,15 @@ import CodingUsQnADetailUI from "./QnADetail.Presenter";
 import {
   CREATE_STACK_COMMENT,
   DELETE_STACK,
+  FETCH_STACK_COMMENT_ORDERBY_CREATE,
+  FETCH_STACK_COMMENT_ORDERBY_LIKE,
   FETCH_STACK_ONE_BY_STACK_ID,
   UPDATE_STACK,
 } from "./QnADetail.Queries";
 
 export default function CodingUsQnADetail() {
   const router = useRouter();
+
   const { data: questionData } = useQuery<
     Pick<IQuery, "fetchStackOnebystackid">,
     IQueryFetchStackOnebystackidArgs
@@ -38,14 +43,22 @@ export default function CodingUsQnADetail() {
     Pick<IMutation, "deleteStack">,
     IMutationDeleteStackArgs
   >(DELETE_STACK);
-  // const { data: answerDataList } = useQuery<
-  //   Pick<IQuery, "fetchAllStackcomment">,
-  //   IQueryFetchAllStackcommentArgs
-  // >(FETCH_ALL_STACK_COMMENT, {variables: {stackid: String(router.query.questionId)}});
+  const { data: answerListDate } = useQuery<
+    Pick<IQuery, "fetchStackCommentorderbycreate">,
+    IQueryFetchStackCommentorderbycreateArgs
+  >(FETCH_STACK_COMMENT_ORDERBY_CREATE, {
+    variables: { stackid: String(router.query.questionId) },
+  });
+  const { data: answerListLike } = useQuery<
+    Pick<IQuery, "fetchStackCommentorderbylike">,
+    IQueryFetchStackCommentorderbylikeArgs
+  >(FETCH_STACK_COMMENT_ORDERBY_LIKE, {
+    variables: { stackid: String(router.query.questionId) },
+  });
 
-  const [isSortGood, setIsSortGood] = useState(true);
+  const [isSortLike, setIsSortLike] = useState(false);
   const toggleSortGubun = () => {
-    setIsSortGood((prev) => !prev);
+    setIsSortLike((prev) => !prev);
   };
   const [editValue, setEditValue] = useState("");
   const onClickButton = (event: MouseEvent<HTMLButtonElement>) => {};
@@ -55,24 +68,26 @@ export default function CodingUsQnADetail() {
       setEditValue(questionData?.fetchStackOnebystackid.contents);
   }, [questionData]);
 
-  const onClickDelete = (id: string) => async () => {
-    // 삭제 확인
+  const onClickDelete =
+    (gubun: "question" | "answer", id: string) => async () => {
+      // 삭제 확인
+      try {
+        if (gubun === "question") {
+          const result = await deleteStack({
+            variables: { stackid: String(router.query.questionId) },
+          });
 
-    try {
-      const result = await deleteStack({
-        variables: { stackid: String(router.query.stackId) },
-      });
-
-      if (!result.data?.deleteStack) {
-        alert("삭제 실패");
-        return;
+          if (!result.data?.deleteStack) {
+            alert("삭제 실패");
+            return;
+          }
+          alert("삭제 성공");
+          router.push("/codingus/question");
+        }
+      } catch (err: any) {
+        // 실패(err.message);
       }
-      alert("삭제 성공");
-      router.push("/codingus/question");
-    } catch (err: any) {
-      // 실패(err.message);
-    }
-  };
+    };
 
   const onClickEditSubmit = (id: string) => async () => {
     try {
@@ -99,7 +114,7 @@ export default function CodingUsQnADetail() {
   const onClickSubmitAnswer = (contents: string) => async () => {
     try {
       const result = await crateStackComment({
-        variables: { contents, stackid: String(router.query.stackId) },
+        variables: { contents, stackid: String(router.query.questionId) },
       });
       if (!result.data?.createStackComment.id) return;
       alert("등록 성공");
@@ -113,33 +128,12 @@ export default function CodingUsQnADetail() {
       children={
         <CodingUsQnADetailUI
           question={questionData?.fetchStackOnebystackid}
-          answers={[
-            {
-              title: "개고숩니다.",
-              contents: "저도 잘 모르겠네요 ㅎㅎㅎㅎ",
-              user: { nickname: "김태훈" },
-              like: 3,
-              dislike: 21,
-              createAt: "2022-02-07T14:42:53.532Z",
-            },
-            {
-              title: "멘토님짱짱",
-              contents: "원두 멘토님께 물어보세요",
-              user: { nickname: "김재민" },
-              like: 13,
-              dislike: 1,
-              createAt: "2022-02-07T14:42:53.532Z",
-            },
-            {
-              title: "집가고싶어요",
-              contents: "흑흑 너무 슬프다",
-              user: { nickname: "최건" },
-              like: 5,
-              dislike: 0,
-              createAt: "2022-02-07T14:42:53.532Z",
-            },
-          ]}
-          isSortGood={isSortGood}
+          answers={
+            isSortLike
+              ? answerListLike?.fetchStackCommentorderbylike
+              : answerListDate?.fetchStackCommentorderbycreate
+          }
+          isSortLike={isSortLike}
           toggleSortGubun={toggleSortGubun}
           onClickButton={onClickButton}
           onClickDelete={onClickDelete}
